+++ Source/AMOBAProject/Public/MOBABaseCharacter.h
@@ -91,7 +91,7 @@ public:
 		int& getMagicResist() { return baseProperty.magicResist; }
 
 	UFUNCTION(BlueprintCallable, Category = "MyMOBA")
-		int& getMoveSpeed() { return baseProperty.moveSpeed; }
+		float& getMoveSpeed() { return baseProperty.moveSpeed; }
 
 	UFUNCTION(BlueprintCallable, Category = "MyMOBA")
 		bool& getBAbleToAttack() { return baseProperty.bAbleToAttack; }
@@ -100,7 +100,7 @@ public:
 		bool& getBIsRemote() { return baseProperty.bIsRemote; }
 
 	UFUNCTION(BlueprintCallable, Category = "MyMOBA")
-		int& getAttackRange() { return baseProperty.attackRange; }
+		float& getAttackRange() { return baseProperty.attackRange; }
 
 	UFUNCTION(BlueprintCallable, Category = "MyMOBA")
 		int& getSeightrange() { return baseProperty.sightRange; }
